Create Circle.py
This is a program that uses the Monte Carlo Method to calculate an approximation of pi. It randomly samples points. It operates under the assumption that the number of points per unit area should be uniform. Therefore, by knowing the ratio of the points inside the circle to the total points, and the ratio of area of the circle to the total area, an approximation of pi can be calculated.