Create DFT.py
This is a program that I created to familiarize myself with the discrete Fourier transform (DFT). I am currently working on some Arduino projects related to analyzing the frequency of sound waves, so the DFT is of particular interest to me. The program takes in data from a CSV and computes the DFT based on the data points. To test the code, I created data for sinusoids of different frequencies in Excel. Not that in line 29, a quantity l is defined. This is essentially the temporal length of the dataset divided by pi. For me, I generated data values at 0.01 second increments.