Create conducting_charge_line.py
I created this code to approximate the charge distribution on a 1D conductor of finite length. Some sources list this as a conducting needle. The principle behind this code is a collection of n finite point charge in static equilibrium. Due to symmetry, the charge distribution should be symmetric about L/2. If each point charge is separated by L / n, a system of linear equations can be solved based on the charges on either side of each charge along with their respective distances, and the constraint that each point charge q0, q1, q2, etc is equal to the total charge Q. The charge density can be approximated by dividing each of the charge values by the value l / n. The endpoint charge density can be calculated by multiplying by 2 l / n. With increasing n, the density approaches a linear distribution in the middle, but the ends turn upwards sharply. The graph looks smooth and continuous.