Create led_beat_change.ino
This code, when connected to an IR LED and a microphone, changes the colors of an IR-controlled LED decoration strip to the beat of music. See https://www.pjrc.com/teensy/td_libs_IRremote.html for more information on decoding IR remote. The beat is detected by increases in the relative power of the audio signal compared to a rolling average. It took a little bit of trial and error to determine the code used by the remote. Different brands of IR receivers/transmitters have different ways of encoding the infrared pulses. The LED set I had used the "NEC" protocol.