Create 2d_heat_equation.py
This is a first draft for a program to show the temperature in a bar with a constant initial temperature, t1 placed into an environment with a constant temperature of t2. Specifically, given the Dirichlet boundary conditions above, I was able to solve the separable heat equation for X(x), Y(y) and T(t). The solution is an infinite sum of each of the 3 eigenfunctions multiplied by their corresponding coefficients. This program is limited to the first 20 terms of the solution.