Create Derivative.py
This is a basic program that plots an approximation of the derivative of a function. After configuring both the domain of the function to be plotted along with the resolution (number of points used), the function can be entered on the line "y =." This program then computes the value of the function at each of the points. With the evaluated values of the function, and the consequences of the mean value theorem, the program estimates the value of the derivative at a point by calculating the slope of a secant line between the points one value higher and one value lower than the point being evaluated. The function is then graphed and the list of approximate derivative values is also plotted. The titles for each subplot can also be adjusted. For this example, the y-axis for the derivative plot was adjusted to better show the function's behavior.