Create force_charge.py
This is code that I made to show the force on charges at different distances from the center of a disk. The function "charge_at_radius()" can be changed to show the force for either a uniformly charged disk (by entering sigma instead of equi_sigma(r)) or a conducting disk at equilibrium by keeping equi_sigma(r).