Add files via upload
This code verifies that the force exerted by a ring with a uniform angular charge density can be calculated by using an elliptic integral. The code also verifies that for all numbers of charges tested, the analytic method produces a numerical value at a fraction of the time taken by the point method.